ObjectDecapsulator tests moved to separated directory and namespace (#2)
