+++ tests/Decapsulator/ObjectDecapsulator/BuildForObjectTest.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Decapsulator;
+namespace Decapsulator\ObjectDecapsulator;
 
-use Decapsulator\AbstractStaticMethodsTest;
+use Decapsulator\ObjectDecapsulator\AbstractStaticMethodsTest;
 use Decapsulator\ObjectDecapsulator;
 
 /**
