+++ tests/Decapsulator/ObjectDecapsulator/PropertyExistsTest.php
@@ -9,12 +9,12 @@
  * file that was distributed with this source code.
  */
 
-namespace Decapsulator;
+namespace Decapsulator\ObjectDecapsulator;
 
-use Decapsulator\AbstractPropertyAccessorsTest;
+use Decapsulator\AbstractPropertyAccessorsTest\ObjectDecapsulator;
 
 /**
- * GetPropertyTest.
+ * PropertyExistsTest.
  * PHPUnit test class for ObjectDecapsulator class.
  *
  * @package Decapsulator
@@ -23,7 +23,7 @@ use Decapsulator\AbstractPropertyAccessorsTest;
  * @license http://http://opensource.org/licenses/MIT MIT License
  * @link http://github.com/exorg/decapsulator
  */
-class GetPropertyTest extends AbstractPropertyAccessorsTest
+class PropertyExistsTest extends AbstractPropertyAccessorsTest
 {
     /**
      * Provide tested method name.
@@ -32,25 +32,35 @@ class GetPropertyTest extends AbstractPropertyAccessorsTest
      */
     protected function provideTestedMethodName()
     {
-        return 'getProperty';
+        return 'propertyExists';
     }
 
     /**
-     * Test getProperty($name) method
-     * gets property value correctly.
+     * Test propertyExists($name) method
+     * returns false when the property does not exist.
+     */
+    public function testReturnsFalseWhenPropertyDoesNotExist()
+    {
+        $arguments = array(self::NONEXISTENT_PROPERTY_NAME);
+
+        $methodReturnedValue = $this->callTestedMethod($arguments);
+
+        $this->assertFalse($methodReturnedValue);
+    }
+
+    /**
+     * Test propertyExists($name) method
+     * returns true when the property exists.
      *
      * @dataProvider existingPropertiesNamesProvider
      * @param string $propertyName
      */
-    public function testGetsPropertyCorrectly($propertyName)
+    public function testReturnsTrueWhenPropertyExists($propertyName)
     {
-        $expectedPropertyValue = 1024;
-        $this->setDecapsulatedObjectProperty($propertyName, $expectedPropertyValue);
-
         $arguments = array($propertyName);
 
-        $actualPropertyValue = $this->callTestedMethod($arguments);
+        $methodReturnedValue = $this->callTestedMethod($arguments);
 
-        $this->assertEquals($expectedPropertyValue, $actualPropertyValue);
+        $this->assertTrue($methodReturnedValue);
     }
 }
